async-io: Allow for multiple simultaneous requests
The current implementation handles all requests in the main event loop,
including blocking operations reading data off of disk. These blocking
operations prevent the server from processing other requests on other
sockets. With this change, leave the main parsing of requests in the
main loop, but move the blocking operations off to a thread from the
main thread pool, significantly increasing throughput of the server.
The thread pool is only enabled when on-disk operations are performed
and not when an in-memory DB is selected.

The number of threads is controlled by the `UV_THREADPOOL_SIZE`
environment variable and defaults to 4.