Write logs asynchronously
Currently, the log function will format the log message and write it
directly to stderr. This is a blocking operation, making in unsuitable
in low latency situations (such as when processing a request).
Re-implement the logging function to queue up messages and flush them to
stderr asynchronously. This is an initial implementation that suffers
from a few problems:

1) It allocates memory, which itself can be blocking
2) No limit is placed on how many messages can be queued up, which can
   be substantial, if data is being written to the log faster that it
   can be flushed out (it may be better to drop entries)