Add logging to most interesting events
Currently, the server only logs a few lines at startup then is totally
silent. Log interesting events to aid in debugging:

- Newly (dis)connected clients
- Any failure interacting with sockets (read error write error)

Include, where possible, the IP and port of the remote end of the
connection, which may help in correlating issues between services.

This is just an initial set of events, which will likely be expanded an
the need arises.