enable TCP keepalive by default
Normally, there is no way to detect whether the remote end of a
connection is still alive. TCP keepalive periodically sends small
packets to the remote end (client in this case) to see if the connection
is still alive. If the remote end does not respond, the connection is
closed, allowing the server to clean up dead connections. This can also
help in environments where network equipment sits between client and
server (e.g., a load balancer) that may tear down connections that have
been idle for too long.

By default, enable TCP keepalive every 60 seconds on all connections
with a command-line option to change the period or disable completely
(set to 0).

Note that this does not prevent idle clients from consuming resources.
To address this, application logic to automatically close connections
that have been idle for too long would be required.